requirements/base.txt
Django==5.0.6
python-environ==0.4.54
psycopg2==2.9.9
Pillow==10.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7